% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cps.R
\name{get_basic_cps}
\alias{get_basic_cps}
\title{Get basic monthly CPS micro data}
\usage{
get_basic_cps(
  year = 2000:2025,
  month = c("jan", "feb", "mar", "apr", "may", "jun", "jul", "aug", "sep", "oct", "nov",
    "dec"),
  variables = NULL,
  weight = c("person", "household", "veteran"),
  state = NULL,
  api_key = Sys.getenv("USCENSUS_KEY")
)
}
\arguments{
\item{year}{Numeric, the year you want to pull the data for}

\item{month}{Character string, the month you want to pull the data for}

\item{variables}{Character string or vector of characters strings of variables IDs. tidycps uses these IDs to pull the micro data.}

\item{weight}{Character string to indicate the type of weight to use in your analysis: person, household or veteran.}

\item{state}{Character string for the state abbreviation or numeric for the state FIPS code. If nothing is entered it pulls the micro data for every state in the United States.}

\item{api_key}{US Census API Key. Recommend to put your census key in .Renviorn for easy access.}
}
\value{
A tibble or tidyverse-ready data frame of CPS micro data.
}
\description{
Get basic monthly CPS micro data
}
\examples{
# Load libraries
pacman::p_load('tidycps',
'tidyverse',
'httr2',
"glue",
"jsonlite",
'ggplot2'
)

# Pull December 2025 basic monthly CPS data for Florida
df <- get_basic_cps(year = 2025, month = 'dec',
weight = "person",
variables = c(
"PRPERTYP", # person type
"PRTAGE", # age
"PESEX", # sex
"PTDTRACE", # race
  "PEHSPNON", # Hispanic
 "PEEDUCA", # educational attainment
  "PEMLR", # labor force status
  "COUNTY"
),
state = 'fl'
)

# Calculate the Florida's December 2025 unemployment rate
fl_labforce_emp <- df |>
  labforce_emp_status() |>
  group_by(state, DATE) |>
  summarize(
    unemployment = sum(unemp*(PWCMPWGT)),
    labor_force = sum(labforce*(PWCMPWGT)),
    unemployment_rate = (unemployment/labor_force)*100
  )
}
