% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cps.R
\name{get_cps_micro_ts}
\alias{get_cps_micro_ts}
\title{Get a Time Series of Basic Monthly CPS micro data}
\usage{
get_cps_micro_ts(
  year = 2000:2025,
  month = 1:12,
  variables = NULL,
  weight = c("person", "household", "veteran"),
  state = NULL,
  api_key = Sys.getenv("USCENSUS_KEY")
)
}
\arguments{
\item{year}{Numeric, the year you want to pull the data for}

\item{month}{Character string, the month you want to pull the data for}

\item{variables}{Character string or vector of characters strings of variables IDs. tidycps uses these IDs to pull the micro data.}

\item{weight}{Character string to indicate the type of weight to use in your analysis: person, household or veteran.}

\item{state}{Character string for the state abbreviation or numeric for the state FIPS code. If nothing is entered it pulls the micro data for every state in the United States.}

\item{api_key}{US Census API Key. Recommend to put your census key in .Renviorn for easy access.}
}
\value{
A tibble or tidyverse-ready data frame of CPS micro data.
}
\description{
Get a Time Series of Basic Monthly CPS micro data
}
