% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cps.R
\name{labforce_emp_status}
\alias{labforce_emp_status}
\title{Create laborforce and employment status indicators}
\usage{
labforce_emp_status(df)
}
\arguments{
\item{df}{A tibble or data frame that contains the variables PRPERTYP, PRTAGE, and PEMLR}
}
\value{
A tibble or data frame with the new variables added
}
\description{
Create laborforce and employment status indicators
}
